Refactor script into modules
